% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{BAF_EM_VAF}
\alias{BAF_EM_VAF}
\title{BAF_EM_VAF}
\usage{
BAF_EM_VAF(MAT, binom = TRUE, show_dots = TRUE, clust_tVAF = NULL)
}
\arguments{
\item{MAT}{A numeric matrix of alternate (AD) and reference (RD) read counts
in addition to pre-calculated total read depth DP = (AD + RD), 
log-transformed binomial coefficient (LBC), log-transformed total 
read depth (log_DP).}

\item{binom}{Boolean set to true by default to model the mixture distribution
assuming binomial distribution. Otherwise set to false to explore beta-binomial
and binomial models.}

\item{show_dots}{Boolean set to true by default to visualize 
computational runtime.}

\item{clust_tVAF}{Boolean but set to null to indicate clustering 
normal read counts represented by modeling a mixture of noise and/or 
VAF cluster centered around 0.5. Otherwise when set to true, the function
models the mixture by noise, cluster at 0.5, clusters p and 1-p for copy-altered 
genomic segments.}
}
\value{
A list of clustered parameter estimates and 
posterior probabilities for inferring classification.
}
\description{
Clusters integer read counts to model mixtures of
noise distributions, binomial vs beta-binomial distributions, 
over-dispersion.
}
