% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_VCFs}
\alias{import_VCFs}
\title{import_VCFs}
\usage{
import_VCFs(DAT, FILTER = NULL, ncores = 1)
}
\arguments{
\item{DAT}{A data.frame containing column names 'FILENAME' for
the full vcf filename and 'STUDYNUMBER' for a unique string
mapped to the control sample used when calling variants 
against the same tumor.}

\item{FILTER}{A list containing named elements 'nDP', 'tDP', 'Qscore',
and 'contigs' for minimum normal depth, minimum tumor depth, minimum
Qscore, and a string of contigs to retain, respectively. Default values
are nDP = tDP = 2, Qscore = 3, and contigs = chr1-chr22, chrX, chrY.}

\item{ncores}{A positive integer for the number of threads to 
use for calculating strand-specific read counts.}
}
\description{
Imports Strelka VCFs containing SNVs and INDELs
}
