% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNMASC.R
\name{new_STRAND}
\alias{new_STRAND}
\title{new_STRAND}
\usage{
new_STRAND(BAM_fn, vcs, minBQ = 13, minMQ = 40, NT = 1)
}
\arguments{
\item{BAM_fn}{A character string for the bam's file path.}

\item{vcs}{A R data.frame of variant calls containing 
columns \code{Chr}, \code{Position}, \code{Ref}, and 
\code{Alt} corresponding to the chromosome,
genomic position, reference base(s), and alternate base(s),
respectively.}

\item{minBQ}{An integer for minimum base quality.}

\item{minMQ}{An integer for minimum mapping quality.}

\item{NT}{A positive integer for the number of threads.}
}
\description{
Extracts strand-specific read counts from a 
bam file.
}
