% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{prep_UNMASC_VCF}
\alias{prep_UNMASC_VCF}
\title{prep_UNMASC_VCF}
\usage{
prep_UNMASC_VCF(outdir, DAT, FILTER, target_fn, anno_fn, ncores = 1)
}
\arguments{
\item{outdir}{A character string of the full path and working directory specifying 
where intermediate and final files are stored.}

\item{DAT}{A data.frame containing column names 'FILENAME' for
the full vcf filename and 'STUDYNUMBER' for a unique string
mapped to the control sample used when calling variants 
against the same tumor.}

\item{FILTER}{A list containing named elements 'nDP', 'tDP', 'Qscore',
and 'contigs' for minimum normal depth, minimum tumor depth, minimum
Qscore, and a string of contigs to retain, respectively. Default values
are nDP = tDP = 2, Qscore = 3, and contigs = chr1-chr22, chrX, chrY.}

\item{target_fn}{A filename for the tab delimited target 
capture bed file. The first three column headers should be 
'Chr', 'Start', and 'End'.}

\item{anno_fn}{A filename for the unique set of variant loci called
across multiple normals and annotated by VEP.}

\item{ncores}{A positive integer for the number of threads to 
use for calculating strand-specific read counts.}
}
\description{
A comprehensive function to import VCFs, annotations,
and target bed file to prepare the necessary VCF for UNMASC's main 
workflow.
}
