% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNMASC.R
\name{run_AF_SEG}
\alias{run_AF_SEG}
\title{run_AF_SEG}
\usage{
run_AF_SEG(
  DATA,
  num_divides,
  init_seg = NULL,
  binom = TRUE,
  clust_tVAF = TRUE,
  show_plot = FALSE,
  show_mess = FALSE,
  min_adj_seg = 20,
  min_nUniqIdx_pSeg = 10
)
}
\arguments{
\item{DATA}{a R data.frame containing columns \code{RD}, 
\code{AD}, and \code{index} corresponding to reference depth,
alternate depth, and the ordering and grouping of \code{DATA}'s 
rows.}

\item{num_divides}{A positive integer to specify the initial 
number of equally spaced segments prior to segmentation. This is
used if \code{init_seg} is set to \code{NULL}.}

\item{init_seg}{A vector specifying the user's initial segmentation.}

\item{binom}{Boolean value set to \code{TRUE} to model read counts
with binomial distribution. Otherwise model read counts with
beta-binomial distribution.}

\item{clust_tVAF}{If \code{TRUE}, \code{DATA} is assumed to 
originate from tumor read counts. Otherwise, \code{DATA} is 
assumed to originate from normal read counts.}

\item{show_plot}{If \code{TRUE}, the function will generate 
a sequence of plots to visualize the points and ongoing segmentation.}

\item{show_mess}{If \code{TRUE}, the function will output detailed 
segmentation messages. By default \code{show_mess = FALSE}.}

\item{min_adj_seg}{An integer for the minimum number of indices
to constitute a segment.}

\item{min_nUniqIdx_pSeg}{An integer for the minimum number of unique
indices per segment.}
}
\description{
This function performs normal and tumor variant
allele frequency segmentation. Rows with allele frequencies 
near 0 and 1 should be excluded before segmentation.
}
