% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_BAF}
\alias{simulate_BAF}
\title{simulate_BAF}
\usage{
simulate_BAF(
  num_segs = 2,
  mean_depth = 300,
  max_PSI = 0,
  tumor = TRUE,
  show_plots = TRUE
)
}
\arguments{
\item{num_segs}{An integer specifying the number of segments to simulate.}

\item{mean_depth}{An integer specifying a mean total depth for loci.}

\item{max_PSI}{The largest beta-binomial overdispersion term a simulated segment can be.}

\item{tumor}{To generate tumor read counts, set to \code{TRUE}, 
otherwise set to \code{FALSE}.}

\item{show_plots}{Set to \code{TRUE} to plot the simulated output.}
}
\value{
A list containing simulated true segmentation and observed data.
}
\description{
This functions simulates B allele frequencies and 
illustrates how UNMASC processes normal and tumor read counts 
for segmentation.
}
