% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\docType{package}
\name{UNMASC-package}
\alias{UNMASC-package}
\title{UNMASC package for tumor-only variant calling with 
unmatched normal controls.}
\description{
UNMASC aims to call somatic mutations without
a matched normal control sample by pooling variant
calls made against unmatched normals. The method
produces data-driven annotation for genomic regions
that may be hard to map to, especially when making
inferences based on the allele frequency. Naturally,
raw tumor-only variant calls contain underlying somatic
point mutations, sample-specific germline variants, and
artifacts. UNMASC attempts to characterize each
variant call through several features to determine 
the underlying variant class.
}
\references{
Little, P., Jo, H., Hoyle, A., Mazul, A.,
Zhao, X., Salazar, A.H., Farquhar, D.,
Sheth, S., Masood, M., Hayward, M.C.,
Parker, J.S., Hoadley, K.A., Zevallos, J.,
Hayes, D.N. (2021)
UNMASC: tumor-only variant calling with 
unmatched normal controls. Nucleic Acids 
Research Cancer, 3. \url{https://doi.org/10.1093/narcan/zcab040}
}
\author{
Paul Little
}
\keyword{package}
