% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNMASC.R
\name{UNMASC_definitions}
\alias{UNMASC_definitions}
\title{UNMASC_definitions}
\usage{
UNMASC_definitions(headers = NULL)
}
\arguments{
\item{headers}{A character vector of column header names. By default
\code{headers = NULL}, in which all available header definitions 
are printed. If a subset of headers are specified, only the subset 
of header definitions are printed.}
}
\value{
Null value, prints definitions to console.
}
\description{
This function prints the column header definitions of 
the output file \code{tumorOnly_VCs.tsv}.
}
